refactor: :recycle: from combat to create Messages
I changed the scope of the result variable, since its refers only to each specific combat.  I removed the part of the combat function that was responsible for creating variables to represent the HTML span elements corresponding to the lives counters, and I moved it to the createMessages function because I think it makes more sense for it to be there since this funtion is responsible for taking the result of the combat and display it on the screen.